planner/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydantic==2.5.0
openai==1.3.0
python-dateutil==2.8.2

